refactor: remove duplicate middlewares folder
