+++ backend/src/routes/admin.js
@@ -2,9 +2,9 @@ const express = require('express');
 const router = express.Router();
 const adminController = require('../controllers/adminController');
 
-const auth = require('../middlewares/authMiddleware');
-const onlyMaster = require('../middlewares/onlyMaster');
-const isAdmin = require('../middlewares/isAdmin');
+const auth = require('../middleware/authMiddleware');
+const onlyMaster = require('../middleware/onlyMaster');
+const isAdmin = require('../middleware/isAdmin');
 
 
 router.get('/users', auth, onlyMaster, adminController.getUsers);
