+++ backend/src/routes/roll.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const rollController = require('../controllers/rollController');
-const auth = require('../middlewares/authMiddleware');
+const auth = require('../middleware/authMiddleware');
 
 router.post('/', auth, rollController.create);
 router.get('/history', auth, rollController.history);
